project directory refactor
The project directory was refactored to follow the _bullet-proof_ react
template. Important components/features were moved into the `feature`
directory, a new `hooks` directory was created to hold custom hooks, and
multiple import statements were updated to adjust for the new custom
paths defined in the `vite.config.ts`.

Basic app functionality remains the same, Cypress test ran on this
commit to ensure the application behaved as expected.
