+++ src/features/DocumentEdit/components/DocumentEdit/DocumentEdit.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import device from '../../utils/device-breakpoints';
+import device from '../../../../utils/device-breakpoints';
 
 const DocumentEditStyled = styled.main`
   grid-area: document;
