+++ src/features/DocumentEdit/components/DocumentEdit/index.ts
@@ -1,0 +1,1 @@
+export { DocumentEdit } from './DocumentEdit';
