+++ src/features/DocumentEdit/components/MarkdownTextArea/index.ts
@@ -1,0 +1,1 @@
+export { MarkdownTextArea } from './MarkdownTextArea';
