+++ src/features/DocumentEdit/components/PreviewTextArea/index.ts
@@ -1,0 +1,1 @@
+export { PreviewTextArea } from './PreviewTextArea';
