+++ src/features/Menu/components/DocumentList/index.ts
@@ -1,0 +1,1 @@
+export { DocumentList } from './DocumentList';
