+++ src/features/TopBar/components/TopBar/index.ts
@@ -1,0 +1,1 @@
+export { TopBar } from './TopBar';
